requirements.txt
# DeliverableEstimate Pro - Dependencies

# Core AI/ML Framework
langgraph>=0.0.40
langchain>=0.1.0
openai>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Excel Processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Environment Management
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.0.0

# Date/Time
python-dateutil>=2.8.0

# Text Processing
beautifulsoup4>=4.12.0

# File Operations
pathlib>=1.0.0

# Testing (Optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Development (Optional)
black>=23.0.0
flake8>=6.0.0